+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Renamed mapDispatch to Dispatch.map()
 - Require render function to Program.run()
 - Make effectManagers optional in Program.run()
+- Remove dep on ts-exhaustive-check
 
 ## [0.1.0](https://github.com/typescript-tea/core/compare/v0.1.0...v0.1.0) - 2020-01-26